refactor: split slack-thread-processor.ts into focused modules
Split the large slack-thread-processor.ts file (1204 lines) into 4 focused modules:

New Modules:
- markdown-to-slack.ts: Markdown to Slack mrkdwn conversion
- blockkit-processor.ts: Blockkit processing and action buttons
- code-block-handler.ts: Code block parsing and validation
- github-actions.ts: GitHub integration and PR buttons

Benefits:
- Reduced main file from 1204 to 693 lines (42% reduction)
- Single responsibility principle applied
- Better maintainability and testability
- Code reusability improved
- Backward compatibility preserved

Fixes #57
