+++ packages/dispatcher/src/queue/slack-thread-processor.ts
@@ -1,527 +1,14 @@
 #!/usr/bin/env bun
 
-import { createHash } from "node:crypto";
 import { WebClient } from "@slack/web-api";
-import { marked } from "marked";
 import PgBoss from "pg-boss";
 import type { GitHubRepositoryManager } from "../github/repository-manager";
-import { generateGitHubAuthUrl } from "../utils/github-utils";
-import { getUserGitHubInfo } from "../slack/handlers/github-handler";
+import { processMarkdownAndBlockkit } from "../converters/blockkit-processor";
+import { generateGitHubActionButtons } from "../converters/github-actions";
+import { convertMarkdownToSlack } from "../converters/markdown-to-slack";
 
-// Generate deterministic action IDs based on content to prevent conflicts during rapid message updates - fixed
-function generateDeterministicActionId(
-  content: string,
-  prefix: string = "action"
-): string {
-  const hash = createHash("sha256")
-    .update(content)
-    .digest("hex")
-    .substring(0, 8);
-  return `${prefix}_${hash}`;
-}
-// Enhanced markdown to Slack conversion with proper handling of all common markdown elements
-function processMarkdownAndBlockkit(content: string): {
-  text: string;
-  blocks: any[];
-} {
-  // Process blockkit with metadata first
-  const codeBlockRegex = /```(\w+)\s*\{([^}]+)\}\s*\n?([\s\S]*?)\n?```/g;
-  let processedContent = content;
-  const actionButtons: any[] = [];
-  let blockIndex = 0; // Track position to ensure unique action_ids
 
-  let match;
-  // biome-ignore lint/suspicious/noAssignInExpressions: Required for regex matching pattern
-  while ((match = codeBlockRegex.exec(content)) !== null) {
-    const [fullMatch, language, metadataStr, codeContent] = match;
 
-    try {
-      const metadata: any = {};
-      metadataStr?.split(",").forEach((pair) => {
-        const [key, value] = pair.split(":").map((s) => s.trim());
-        if (key && value) {
-          const cleanKey = key.replace(/"/g, "");
-          let cleanValue: any = value.replace(/"/g, "");
-          if (cleanValue === "true") cleanValue = true;
-          if (cleanValue === "false") cleanValue = false;
-          metadata[cleanKey] = cleanValue;
-        }
-      });
-
-      if (metadata.action) {
-        logger.debug(
-          `Found action block - language: ${language}, action: ${metadata.action}, show: ${metadata.show}`
-        );
-
-        if (language === "blockkit") {
-          // Always hide the code block from the message for blockkit actions
-          // The show parameter doesn't affect button creation for blockkit
-          processedContent = processedContent.replace(fullMatch, "");
-
-          const parsed = codeContent
-            ? JSON.parse(codeContent.trim())
-            : { blocks: [] };
-          const buttonValue = JSON.stringify({
-            blocks: parsed.blocks || [parsed],
-          });
-
-          // Skip the button entirely if value exceeds 2000 chars (Slack limit)
-          if (buttonValue.length > 2000) {
-            console.log(
-              `[DEBUG] Skipping blockkit button - exceeds 2000 char limit (${buttonValue.length} chars), action: ${metadata.action}`
-            );
-            continue;
-          }
-
-          const actionId = generateDeterministicActionId(
-            codeContent + metadata.action + blockIndex,
-            "blockkit_form"
-          );
-          const button = {
-            type: "button",
-            text: { type: "plain_text", text: metadata.action },
-            action_id: actionId,
-            value: buttonValue,
-          };
-          actionButtons.push(button);
-          console.log(
-            `[DEBUG] Added blockkit button - action: ${metadata.action}, actionId: ${actionId}`
-          );
-        } else {
-          // For non-blockkit actions (bash, python, etc.)
-          // Hide the code block unless show: true
-          if (metadata.show !== true) {
-            processedContent = processedContent.replace(fullMatch, "");
-          }
-
-          // Skip entirely if show: false (no button)
-          if (metadata.show === false) {
-            console.log(
-              `[DEBUG] Skipping ${language} action with show:false - action: ${metadata.action}`
-            );
-            continue;
-          }
-
-          if (codeContent) {
-            // Skip the button entirely if value exceeds 2000 chars (Slack limit)
-            if (codeContent.length > 2000) {
-              console.log(
-                `[DEBUG] Skipping ${language} button - exceeds 2000 char limit (${codeContent.length} chars), action: ${metadata.action}`
-              );
-              continue;
-            }
-
-            const actionId = generateDeterministicActionId(
-              codeContent + metadata.action + blockIndex,
-              language
-            );
-            const button = {
-              type: "button",
-              text: { type: "plain_text", text: metadata.action },
-              action_id: `${language}_${actionId}`,
-              value: codeContent,
-            };
-            actionButtons.push(button);
-            console.log(
-              `[DEBUG] Added ${language} button - action: ${metadata.action}, actionId: ${language}_${actionId}`
-            );
-          }
-        }
-      }
-
-      blockIndex++; // Increment for each processed block to ensure unique action_ids
-    } catch (error) {
-      console.error("Failed to parse code block:", error);
-    }
-  }
-
-  // Enhanced markdown to Slack conversion
-  const text = convertMarkdownToSlack(processedContent);
-
-  // Always create at least one block
-  const blocks: any[] = [];
-
-  if (text) {
-    // Slack has a 3000 character limit for text in section blocks
-    const MAX_TEXT_LENGTH = 3000;
-
-    if (text.length <= MAX_TEXT_LENGTH) {
-      blocks.push({
-        type: "section",
-        text: {
-          type: "mrkdwn",
-          text: text,
-        },
-      });
-    } else {
-      // Split long text into multiple blocks
-      let remainingText = text;
-      while (remainingText.length > 0) {
-        // Take up to MAX_TEXT_LENGTH characters, but try to break at a newline if possible
-        let chunk = remainingText.substring(0, MAX_TEXT_LENGTH);
-
-        // If we're not at the end and we're cutting mid-text, try to find a better break point
-        if (remainingText.length > MAX_TEXT_LENGTH) {
-          const lastNewline = chunk.lastIndexOf("\n");
-          if (lastNewline > MAX_TEXT_LENGTH * 0.8) {
-            // If there's a newline in the last 20% of the chunk
-            chunk = chunk.substring(0, lastNewline);
-          }
-        }
-
-        blocks.push({
-          type: "section",
-          text: {
-            type: "mrkdwn",
-            text: chunk,
-          },
-        });
-
-        remainingText = remainingText.substring(chunk.length).trim();
-      }
-    }
-  }
-
-  if (actionButtons.length > 0) {
-    console.log(
-      `[DEBUG] Adding ${actionButtons.length} action buttons to blocks`
-    );
-    if (blocks.length > 0) blocks.push({ type: "divider" });
-    blocks.push({
-      type: "actions",
-      elements: actionButtons,
-    });
-  }
-
-  console.log(
-    `[DEBUG] processMarkdownAndBlockkit returning - text length: ${text?.length || 0}, blocks count: ${blocks.length}, action buttons: ${actionButtons.length}`
-  );
-
-  return { text, blocks };
-}
-
-/**
- * Custom renderer for converting markdown to Slack's mrkdwn format
- */
-class SlackRenderer extends marked.Renderer {
-  heading(text: string, _level: number): string {
-    // Convert headings - preserve inline formatting like bold/italic
-    // Headers themselves are not automatically bold in Slack
-
-    let processedText = text;
-
-    // Convert markdown bold (**text**) to Slack bold (*text*)
-    processedText = processedText.replace(/\*\*(.+?)\*\*/g, "*$1*");
-
-    // Convert markdown bold (__text__) to Slack bold (*text*)
-    processedText = processedText.replace(/__(.+?)__/g, "*$1*");
-
-    // Convert markdown italic (*text*) to Slack italic (_text_)
-    // But be careful not to convert Slack bold markers we just added
-    processedText = processedText.replace(
-      /(?<!\*)\*(?!\*)([^*]+)(?<!\*)\*(?!\*)/g,
-      "_$1_"
-    );
-
-    // Add extra spacing after headers for visual separation
-    return `${processedText}\n\n`;
-  }
-
-  paragraph(text: string): string {
-    return `${text}\n\n`;
-  }
-
-  strong(text: string): string {
-    // Bold in Slack is *text*
-    return `*${text}*`;
-  }
-
-  em(text: string): string {
-    // Italic in Slack is _text_
-    return `_${text}_`;
-  }
-
-  code(text: string): string {
-    // Inline code in Slack is `text`
-    return `\`${text}\``;
-  }
-
-  codespan(text: string): string {
-    // Inline code in Slack is `text`
-    return `\`${text}\``;
-  }
-
-  blockquote(quote: string): string {
-    // Convert blockquote to italic with quote prefix
-    const lines = quote.trim().split("\n");
-    return `${lines.map((line) => `_> ${line.trim()}_`).join("\n")}\n\n`;
-  }
-
-  list(body: string, _ordered: boolean, _start: number | ""): string {
-    return `${body}\n`;
-  }
-
-  listitem(text: string, _task?: boolean, _checked?: boolean): string {
-    // Slack supports bullet points and numbered lists
-    return `• ${text.trim()}\n`;
-  }
-
-  link(href: string, _title: string | null | undefined, text: string): string {
-    // Slack link format is <url|text>
-    return `<${href}|${text}>`;
-  }
-
-  br(): string {
-    return "\n";
-  }
-
-  hr(): string {
-    return "\n---\n\n";
-  }
-}
-
-/**
- * Convert markdown to Slack's mrkdwn format using marked with custom renderer
- */
-export function convertMarkdownToSlack(content: string): string {
-  const renderer = new SlackRenderer();
-
-  // First, handle raw triple backtick code blocks that might not be properly formatted
-  // This handles cases where content has ```language\ncode\n``` format
-  const preprocessed = content.replace(
-    /```(\w*)\n?([\s\S]*?)```/g,
-    (match, lang, code) => {
-      // Convert to a format that marked can handle properly
-      if (code?.trim()) {
-        // Use HTML pre/code tags that marked will process
-        const langAttr = lang ? ` class="language-${lang}"` : "";
-        return `<pre><code${langAttr}>${code.trim()}</code></pre>`;
-      }
-      return match;
-    }
-  );
-
-  // Configure marked options
-  marked.setOptions({
-    renderer: renderer,
-    breaks: true, // Convert single line breaks to <br>
-    gfm: true, // GitHub flavored markdown
-  });
-
-  try {
-    let processed = marked.parse(preprocessed) as string;
-
-    // Clean up extra whitespace but preserve intentional line breaks
-    processed = processed
-      .replace(/\n{3,}/g, "\n\n") // Max 2 consecutive newlines
-      .trim();
-
-    // Handle code blocks specially - marked converts them to HTML, we need to convert back to Slack format
-    // Note: Slack doesn't support triple backtick code blocks in text fields, only in blocks
-    // So we'll convert code blocks to single-line code format for the text field
-    processed = processed.replace(
-      /<pre><code(?:\s+class="language-(\w+)")?>([\s\S]*?)<\/code><\/pre>/g,
-      (_match, _language, code) => {
-        // Decode HTML entities in code blocks
-        const decodedCode = code
-          .replace(/&lt;/g, "<")
-          .replace(/&gt;/g, ">")
-          .replace(/&amp;/g, "&")
-          .replace(/&quot;/g, '"')
-          .replace(/&#39;/g, "'");
-
-        // For Slack text field, use single backticks for inline code
-        // For multi-line code, we'll use indentation instead of backticks
-        // since Slack text fields don't support proper code blocks
-        const lines = decodedCode.trim().split("\n");
-        if (lines.length === 1) {
-          return `\`${lines[0]}\``;
-        } else {
-          // For multi-line code, use indentation (4 spaces) instead of backticks
-          // This preserves the code structure without causing issues with # symbols
-          return lines.map((line: string) => `    ${line}`).join("\n");
-        }
-      }
-    );
-
-    // Clean up any remaining HTML entities that might have been introduced
-    processed = processed
-      .replace(/&gt;/g, ">")
-      .replace(/&lt;/g, "<")
-      .replace(/&amp;/g, "&")
-      .replace(/&quot;/g, '"')
-      .replace(/&#39;/g, "'");
-
-    return processed;
-  } catch (error) {
-    console.error("Failed to parse markdown:", error);
-    // Fallback to original content if parsing fails
-    return content;
-  }
-}
-
-/**
- * Generate GitHub action buttons for the session branch
- */
-async function generateGitHubActionButtons(
-  userId: string,
-  gitBranch: string | undefined,
-  hasGitChanges: boolean | undefined,
-  pullRequestUrl: string | undefined,
-  userMappings: Map<string, string>,
-  repoManager: GitHubRepositoryManager,
-  slackClient?: any
-): Promise<any[] | undefined> {
-  try {
-    logger.debug(
-      `Generating GitHub action buttons for user ${userId}, gitBranch: ${gitBranch}, hasGitChanges: ${hasGitChanges}, pullRequestUrl: ${pullRequestUrl}`
-    );
-
-    // If no git branch provided, don't show buttons
-    if (!gitBranch) {
-      logger.debug(`No git branch provided, skipping GitHub buttons`);
-      return undefined;
-    }
-
-    // Check if we're on a session branch (indicates work has been done)
-    const isSessionBranch = gitBranch.startsWith("claude/");
-
-    // Show buttons if:
-    // 1. There are uncommitted changes, OR
-    // 2. An existing PR exists, OR
-    // 3. We're on a session branch (even if all changes are committed)
-    if (!hasGitChanges && !pullRequestUrl && !isSessionBranch) {
-      logger.debug(
-        `No git changes, no PR, and not a session branch, skipping GitHub buttons`
-      );
-      return undefined;
-    }
-
-    // Get GitHub username from Slack user ID
-    let githubUsername = userMappings.get(userId);
-    if (!githubUsername && slackClient) {
-      // Create user mapping on-demand if not found
-      logger.debug(`Creating on-demand user mapping for user ${userId}`);
-      try {
-        const userInfo = await slackClient.users.info({ user: userId });
-        const user = userInfo.user;
-
-        let username =
-          user.profile?.display_name || user.profile?.real_name || user.name;
-        if (!username) {
-          username = userId;
-        }
-
-        // Sanitize username for GitHub
-        username = username
-          .toLowerCase()
-          .replace(/[^a-z0-9-]/g, "-")
-          .replace(/^-|-$/g, "");
-
-        username = `user-${username}`;
-        userMappings.set(userId, username);
-        githubUsername = username;
-
-        logger.info(`Created user mapping: ${userId} -> ${username}`);
-      } catch (error) {
-        logger.error(`Failed to create user mapping for ${userId}:`, error);
-        const fallbackUsername = `user-${userId.substring(0, 8)}`;
-        userMappings.set(userId, fallbackUsername);
-        githubUsername = fallbackUsername;
-      }
-    }
-
-    if (!githubUsername) {
-      logger.debug(`No GitHub username mapping found for user ${userId}`);
-      return undefined;
-    }
-
-    // Get repository information, create if needed
-    const repository = await repoManager.ensureUserRepository(githubUsername);
-    if (!repository) {
-      logger.debug(`No repository found for GitHub user ${githubUsername}`);
-      return undefined;
-    }
-
-    const repoUrl = repository.repositoryUrl;
-    const repoPath = repoUrl.replace("https://github.com/", "");
-
-    logger.info(
-      `Showing action buttons for branch: ${gitBranch}, PR exists: ${!!pullRequestUrl}`
-    );
-
-    // Check if user has GitHub authentication
-    const githubInfo = await getUserGitHubInfo(userId);
-    const hasGitHubAuth = !!githubInfo.token;
-
-    // Generate auth URL for GitHub OAuth
-    const authUrl = generateGitHubAuthUrl(userId);
-
-    const buttons: any[] = [];
-
-    // Show appropriate PR button based on whether PR exists
-    if (pullRequestUrl) {
-      // PR exists - show view button with green checkmark
-      buttons.push({
-        type: "button",
-        text: { type: "plain_text", text: "🔀 View Pull Request" },
-        url: pullRequestUrl,
-        action_id: generateDeterministicActionId(
-          `view_pr_${repoPath}_${gitBranch}`,
-          "github_view_pr"
-        ),
-      });
-    } else if (hasGitChanges || isSessionBranch) {
-      // Build dynamic prompt based on GitHub auth status
-      let prompt: string;
-
-      if (hasGitHubAuth) {
-        // User is authenticated - agent should handle forking if needed
-        prompt = `📝 *Create Pull Request*
-
-• Review the code and cleanup any temporary files
-• Commit all changes to Git
-• Try to push: \`git push -u origin ${gitBranch}\`
-• If push fails due to permissions:
-  - Fork the repository: \`gh repo fork --clone=false\`
-  - Add fork as remote: \`git remote add fork $(gh repo fork --clone=false 2>&1 | grep -o 'https://[^ ]*')\`
-  - Push to fork: \`git push -u fork ${gitBranch}\`
-• Create PR: \`gh pr create --web\``;
-      } else {
-        // User is not authenticated - provide auth link
-        prompt = `📝 *Create Pull Request*
-
-• Review the code and cleanup any temporary files
-• Commit all changes to Git
-• Try to push: \`git push -u origin ${gitBranch}\`
-• If you get a permission error (403), ask the user to authenticate [here](${authUrl})
-• After authentication, retry the push
-• Create PR: \`gh pr create --web\``;
-      }
-
-      // No PR but has changes OR on a session branch - show create PR button
-      buttons.push({
-        type: "button",
-        text: { type: "plain_text", text: "🔀 Create Pull Request" },
-        action_id: generateDeterministicActionId(
-          `pr_${repoPath}_${gitBranch}`,
-          "github_pr"
-        ),
-        value: JSON.stringify({
-          action: "create_pr",
-          repo: repoPath,
-          branch: gitBranch,
-          prompt: prompt,
-        }),
-      });
-    }
-
-    return buttons.length > 0 ? buttons : undefined;
-  } catch (_error) {
-    // Return undefined on error - this will result in no action buttons being added
-    return undefined;
-  }
-}
 
 import logger from "../logger";
 
@@ -1201,3 +688,6 @@ export class ThreadResponseConsumer {
     };
   }
 }
+
+// Export functions for backward compatibility
+export { convertMarkdownToSlack };
